Refactor search files
