+++ src/lib/components/documents/search/searchFields.ts
