+++ src/lib/components/documents/search/tests/searchFields.test.ts
