+++ src/lib/components/documents/search/types.d.ts
@@ -1,0 +1,9 @@
+import type { FilterFields } from "../Filter.svelte";
+import type { SortDirection, SortField } from "../Sort.svelte";
+
+export interface SearchProps {
+  query?: string;
+  filters?: FilterFields;
+  sort?: SortField;
+  direction?: SortDirection;
+}
