+++ src/lib/components/layouts/DocumentListToolbar.svelte
@@ -5,7 +5,7 @@
   import Dropdown from "../common/Dropdown.svelte";
   import Menu from "../common/Menu.svelte";
   import PageToolbar from "../common/PageToolbar.svelte";
-  import DocumentSearch from "../documents/Search.svelte";
+  import DocumentSearch from "../documents/search/Search.svelte";
   import SidebarItem from "../sidebar/SidebarItem.svelte";
   import VisibleFields from "../documents/VisibleFields.svelte";
